requirements.txt
openai>=1.0.0
requests==2.31.0
crewai>=0.32.0
langchain>=0.1.0
langchain_community>=0.0.0
langchain_openai>=0.0.0
litellm>=1.0.0
Flask>=2.3.0
Werkzeug>=2.3.0
redis>=5.0.0
python-dotenv>=1.0.0
streamlit